feat(solver): draft for fuse relaxed and reactive-limits solvers
