+++ src/main/java/com/powsybl/openloadflow/knitro/solver/AbstractKnitroSolver.java
@@ -33,7 +33,7 @@ import static com.google.common.primitives.Doubles.toArray;
  *      - Configuration of the external Knitro solver.
  *      - Creation of the optimization problem (extending {@link AbstractKnitroProblem}).
  *      - Processing of the obtained solution.
- * This class can be extended to add custom behavior to any of these features (e.g., in {@link RelaxedKnitroSolver}).
+ * This class can be extended to add custom behavior to any of these features (e.g., in {@link AbstractRelaxedKnitroSolver}).
  * For example, if you modify the optimization problem, you may also need to update the solution-processing logic.
  *
  * @author Pierre Arvy {@literal <pierre.arvy at artelys.com>}
