+++ src/main/java/com/powsybl/openloadflow/knitro/solver/KnitroCallbacks.java
@@ -113,7 +113,7 @@ public final class KnitroCallbacks {
          * Allows modification of the constraint value.
          * Default implementation returns no modification.
          * Should be overridden by subclasses that modify the callbacks of Jacobian matrix
-         * (e.g., to add relaxation variables in {@link RelaxedKnitroSolver.RelaxedKnitroProblem}).
+         * (e.g., to add relaxation variables in {@link AbstractRelaxedKnitroSolver.RelaxedKnitroProblem}).
          *
          * @param equationId The equation ID.
          * @param equationType The equation type.
@@ -247,7 +247,7 @@ public final class KnitroCallbacks {
         /**
          * Computes a modified Jacobian value. Default implementation returns 0.
          * Should be overridden by subclasses that modify the callbacks of Jacobian matrix
-         * (e.g., to add relaxation variables in {@link RelaxedKnitroSolver.RelaxedKnitroProblem}).
+         * (e.g., to add relaxation variables in {@link AbstractRelaxedKnitroSolver.RelaxedKnitroProblem}).
          */
         protected double computeModifiedJacobianValue(int variableIndex, int constraintIndex) {
             return 0.0;
